+++ lib/pigeon/fcm.ex
@@ -107,15 +107,19 @@ defmodule Pigeon.FCM do
   @max_retries 3
 
   defstruct config: nil,
-            queue: Pigeon.NotificationQueue.new(),
+            queue: Pigeon.HTTP.RequestQueue.new(),
             retries: @max_retries,
-            socket: nil,
-            stream_id: 1
+            socket: nil
 
   @behaviour Pigeon.Adapter
 
-  alias Pigeon.{Configurable, NotificationQueue}
-  alias Pigeon.Http2.{Client, Stream}
+  import Pigeon.Tasks, only: [process_on_response: 1]
+
+  alias Pigeon.Configurable
+  alias Pigeon.FCM.Error
+  alias Pigeon.HTTP.RequestQueue
+
+  require Logger
 
   @impl Pigeon.Adapter
   def init(opts) do
@@ -137,41 +141,38 @@ defmodule Pigeon.FCM do
 
   @impl Pigeon.Adapter
   def handle_push(notification, state) do
-    %{config: config, queue: queue} = state
+    %{config: config, queue: queue, socket: socket} = state
     headers = Configurable.push_headers(config, notification, [])
     payload = Configurable.push_payload(config, notification, [])
+    method = "POST"
+    path = "/v1/projects/#{config.project_id}/messages:send"
 
-    Client.default().send_request(state.socket, headers, payload)
+    {:ok, socket, ref} =
+      Mint.HTTP.request(socket, method, path, headers, payload)
 
-    new_q = NotificationQueue.add(queue, state.stream_id, notification)
+    new_q = RequestQueue.add(queue, ref, notification)
 
     state =
       state
-      |> inc_stream_id()
+      |> Map.put(:socket, socket)
       |> Map.put(:queue, new_q)
 
     {:noreply, state}
   end
 
   @impl Pigeon.Adapter
-  def handle_info(:ping, state) do
-    Client.default().send_ping(state.socket)
-    Configurable.schedule_ping(state.config)
+  def handle_info(:ping, %{config: config, socket: socket} = state) do
+    {:ok, socket, _ref} = Mint.HTTP2.ping(socket)
+    Configurable.schedule_ping(config)
 
-    {:noreply, state}
+    {:noreply, %{state | socket: socket}}
   end
 
   def handle_info({:closed, _}, %{config: config} = state) do
     case connect_socket(config) do
       {:ok, socket} ->
         Configurable.schedule_ping(config)
-
-        state =
-          state
-          |> reset_stream_id()
-          |> Map.put(:socket, socket)
-
-        {:noreply, state}
+        {:noreply, %{state | socket: socket}}
 
       {:error, reason} ->
         {:stop, reason}
@@ -179,9 +180,45 @@ defmodule Pigeon.FCM do
   end
 
   def handle_info(msg, state) do
-    case Client.default().handle_end_stream(msg, state) do
-      {:ok, %Stream{} = stream} -> process_end_stream(stream, state)
-      _else -> {:noreply, state}
+    %{queue: queue, socket: socket} = state
+
+    case Mint.HTTP.stream(socket, msg) do
+      :unknown ->
+        {:noreply, state}
+
+      {:ok, socket, responses} ->
+        {done, new_q} =
+          responses
+          |> RequestQueue.process(queue)
+          |> RequestQueue.pop_done()
+
+        for {_ref, request} <- done do
+          handle_response(request)
+        end
+
+        {:noreply, %{state | queue: new_q, socket: socket}}
+
+      {:error, socket, error, _responses} ->
+        error |> inspect(pretty: true) |> Logger.error()
+        {:noreply, %{state | socket: socket}}
+    end
+  end
+
+  def handle_response(%{body: body, notification: notif}) do
+    body
+    |> Pigeon.json_library().decode!()
+    |> case do
+      %{"name" => name} ->
+        notif
+        |> Map.put(:name, name)
+        |> Map.put(:response, :success)
+        |> process_on_response()
+
+      %{"error" => error} ->
+        notif
+        |> Map.put(:error, error)
+        |> Map.put(:response, Error.parse(error))
+        |> process_on_response()
     end
   end
 
@@ -199,30 +236,5 @@ defmodule Pigeon.FCM do
           {:error, reason}
         end
     end
-  end
-
-  @doc false
-  def process_end_stream(%Stream{id: stream_id} = stream, state) do
-    %{queue: queue, config: config} = state
-
-    case NotificationQueue.pop(queue, stream_id) do
-      {nil, new_queue} ->
-        # Do nothing if no queued item for stream
-        {:noreply, %{state | queue: new_queue}}
-
-      {notif, new_queue} ->
-        Configurable.handle_end_stream(config, stream, notif)
-        {:noreply, %{state | queue: new_queue}}
-    end
-  end
-
-  @doc false
-  def inc_stream_id(%{stream_id: stream_id} = state) do
-    %{state | stream_id: stream_id + 2}
-  end
-
-  @doc false
-  def reset_stream_id(state) do
-    %{state | stream_id: 1}
   end
 end