+++ test/pigeon/http2_client_test.exs
@@ -1,4 +1,0 @@
-defmodule Pigeon.Http2.ClientTest do
-  use ExUnit.Case
-  doctest Pigeon.Http2.Client.Kadabra
-end